web-interface/requirements.txt
absl-py==1.0.0
astunparse==1.6.3
attrs==19.3.0
Automat==0.8.0
bidict==0.22.0
blinker==1.4
cachetools==5.2.0
celluloid==0.2.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
cloud-init==22.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
deepdish==0.3.7
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
easing-functions==1.0.4
entrypoints==0.3
Flask==2.1.2
Flask-SocketIO==5.2.0
flatbuffers==1.12
gast==0.4.0
google-auth==2.9.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
gunicorn==20.1.0
h5py==3.7.0
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
imageio==2.19.3
importlib-metadata==4.12.0
incremental==16.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keras==2.9.0
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libclang==14.0.1
lxml==4.9.1
macaroonbakery==1.3.1
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.4.3
mediapipe==0.8.10
more-itertools==4.2.0
netifaces==0.10.4
networkx==2.8.4
nibabel==4.0.1
nilearn==0.9.1
nltools==0.4.5
numexpr==2.8.3
numpy==1.21.2
oauthlib==3.1.0
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
pexpect==4.6.0
Pillow==8.3.2
protobuf==3.19.4
py-feat==0.4.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pynv==0.2
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-engineio==4.3.3
python-socketio==5.7.0
pytz==2022.1
PyWavelets==1.3.0
PyYAML==5.3.1
redis==3.3.11
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rq==1.2.2
rsa==4.8
scikit-image==0.19.3
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sos==4.3
ssh-import-id==5.10
systemd-python==234
tables==3.7.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2022.5.4
torch==1.12.0
torchvision==0.13.0
tqdm==4.64.0
Twisted==18.9.0
typing-extensions==4.3.0
ubuntu-advantage-tools==27.9
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
Werkzeug==2.1.2
wrapt==1.14.1
zipp==1.0.0
zope.interface==4.7.1
